chore(deps): Update project dependencies
Upgrades several key packages to their latest versions, including `boto3`, `psycopg`, and `django-storages`. This brings in the latest features, bug fixes,and security patches.

Additionally, this commit removes the redundant `psycopg-binary` and `psycopg2-binary` packages, as `psycopg` (v3) is the only required adapter for PostgreSQL in this project. This simplifies the dependency tree and resolves potential conflicts.